Safety checks, refactoring still
More safety checks for moving to the mission giver
Added range checks to make sure everyone is in range before trying to zone in, or before Invising if needed
Make sure all other toons actually zone in before the driver zones in
More checks to make sure everyone actually zoned in
Zone checks while the script is running in case the driver is no longer in the mission zone
Started work on making the CWTN calls configuration specific, still a WIP
Refactored the Invis routine to only single invis if doube-invis is not needed
Added some more checks to make sure where the driver is in the zone if the script is started in-zone. To make sure he does not run back to the starting point before or after triggering the event.
Increased the timeout for waiting for someone to zone in from 1 minute to 10 minutes.  To give someone time to get back form AFK and zone someone in if needed.
Added Burn flag to the ini / config section at the top of the init script
Stores the CWTN mode for the driver before starting and sets it back at the end of the script.